refactored line processing to make it simpler and actually understandable
